refactor(core/stellar): move SEP-41 token resolution to tokens.py.
sac_address_from_asset reaches into writers.py, which reaches back for the
strkey primitives -- a cycle that held together only because the import was
deferred into the function body. A separate module breaks it, leaves helpers.py
as plain strkey primitives, and gives the token code a home that layout.py can
import from.

[no changelog]
