+++ core/embed/upymod/qstrdefsport.h
@@ -662,6 +662,7 @@ Q(apps.stellar.operations.layout)
 Q(apps.stellar.operations.serialize)
 Q(apps.stellar.sign_soroban_authorization)
 Q(apps.stellar.sign_tx)
+Q(apps.stellar.tokens)
 Q(apps.stellar.writers)
 Q(apps.tezos)
 Q(apps.tezos.get_address)
